Address review comments, Deduplicate get_thread_id
