+++ subscriber/src/future/subscriber.rs
@@ -3,10 +3,8 @@ use tracing_core::Event;
 use tracing_core::Subscriber;
 use tracing_core::{Interest, Metadata};
 
-use std::cell::{Cell, RefCell};
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::{Arc, Once, RwLock};
-use std::thread;
+use std::sync::atomic::Ordering;
+use std::sync::{Arc, RwLock};
 
 use chrono::prelude::*;
 
@@ -16,36 +14,12 @@ use crate::messages::listen_response::Variant;
 use crate::messages::Recorder;
 use crate::*;
 
-static THREAD_COUNTER: AtomicUsize = AtomicUsize::new(1);
-
-thread_local! {
-    static THREAD_ID_INIT: Once = Once::new();
-    static THREAD_ID: Cell<usize> = Cell::new(0);
-
-    static STACK: RefCell<Vec<SpanId>> = RefCell::new(Vec::new());
-}
-
-fn get_thread_id(console: &ConsoleForwarder) -> ThreadId {
-    THREAD_ID_INIT.with(|init_guard| {
-        init_guard.call_once(|| {
-            THREAD_ID.with(|id| {
-                let new_id = THREAD_COUNTER.fetch_add(1, Ordering::SeqCst);
-                if let Some(name) = thread::current().name() {
-                    console.register_thread_name(ThreadId(new_id), name.to_string());
-                }
-                id.set(new_id);
-            })
-        });
-        THREAD_ID.with(|id| ThreadId(id.get()))
-    })
-}
-
 pub struct ConsoleForwarder {
     pub(crate) tx: UnboundedSender<Variant>,
     pub(crate) registry: Arc<RwLock<crate::future::server::Registry>>,
 }
 
-impl ConsoleForwarder {
+impl ThreadNameRegister for ConsoleForwarder {
     fn register_thread_name(&self, id: ThreadId, name: String) {
         self.registry.write().unwrap().thread_names.insert(id, name);
     }
@@ -55,6 +29,7 @@ impl Subscriber for ConsoleForwarder {
     fn enabled(&self, _metadata: &Metadata) -> bool {
         true
     }
+
     fn new_span(&self, span: &span::Attributes) -> span::Id {
         let id = self.registry.write().unwrap().new_id();
         let mut rec = Recorder::default();
@@ -73,6 +48,7 @@ impl Subscriber for ConsoleForwarder {
 
         id.as_span()
     }
+
     fn record(&self, span: &span::Id, values: &span::Record) {
         let mut recorder = messages::Recorder::default();
         values.record(&mut recorder);
@@ -88,6 +64,7 @@ impl Subscriber for ConsoleForwarder {
             }))
             .expect("Channel not closed");
     }
+
     fn record_follows_from(&self, span: &span::Id, follows: &span::Id) {
         self.tx
             .unbounded_send(Variant::Follows(messages::RecordFollowsFrom {
@@ -96,6 +73,7 @@ impl Subscriber for ConsoleForwarder {
             }))
             .expect("Channel not closed");
     }
+
     fn event(&self, event: &Event) {
         let mut recorder = messages::Recorder::default();
         event.record(&mut recorder);
@@ -109,7 +87,7 @@ impl Subscriber for ConsoleForwarder {
             is_contextual: event.is_contextual(),
             is_root: event.is_root(),
             metadata: Some(event.metadata().into()),
-            parent: event.parent().map(|p| p.into()),
+            parent: event.parent().map(Into::into),
         };
 
         self.tx
@@ -125,18 +103,22 @@ impl Subscriber for ConsoleForwarder {
             }))
             .expect("Channel not closed");
     }
+
     fn enter(&self, span: &span::Id) {
         STACK.with(|stack| stack.borrow_mut().push(SpanId::new(span.into_u64())))
     }
+
     fn exit(&self, _: &span::Id) {
         STACK.with(|stack| stack.borrow_mut().pop());
     }
+
     fn register_callsite(&self, metadata: &'static Metadata<'static>) -> Interest {
         match self.enabled(metadata) {
             true => Interest::always(),
             false => Interest::never(),
         }
     }
+
     fn clone_span(&self, id: &span::Id) -> span::Id {
         let index = SpanId::new(id.into_u64()).as_index();
         self.registry.read().unwrap().spans[index]
@@ -146,6 +128,7 @@ impl Subscriber for ConsoleForwarder {
             .fetch_add(1, Ordering::SeqCst);
         id.clone()
     }
+
     fn drop_span(&self, ref id: span::Id) {
         let span_id = SpanId::new(id.into_u64());
         let index = span_id.as_index();
@@ -156,8 +139,8 @@ impl Subscriber for ConsoleForwarder {
             .fetch_sub(1, Ordering::SeqCst);
         if old_count == 1 {
             let mut registry = try_lock!(self.registry.write());
-            let next_id = std::mem::replace(&mut registry.next_id, Some(span_id));
-            std::mem::replace(&mut registry.spans[index], SpanState::Free { next_id });
+            let next_id = registry.next_id.replace(span_id);
+            registry.spans[index] = SpanState::Free { next_id };
         }
     }
 }
